+++ src/main/java/potatoes/server/travel/service/TravelService.java
@@ -1,19 +1,13 @@
 package potatoes.server.travel.service;
 
 import static potatoes.server.utils.error.ErrorCode.*;
-import static potatoes.server.utils.time.DateTimeUtils.*;
 
 import java.time.Duration;
-import java.time.Instant;
 import java.time.ZoneOffset;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -25,12 +19,10 @@ import potatoes.server.infra.s3.S3UtilsProvider;
 import potatoes.server.travel.bookmark.entity.Bookmark;
 import potatoes.server.travel.bookmark.repository.BookmarkRepository;
 import potatoes.server.travel.dto.CreateTravelRequest;
-import potatoes.server.travel.dto.GetMyTravelResponse;
 import potatoes.server.travel.dto.ParticipantResponse;
 import potatoes.server.travel.dto.SimpleTravelResponse;
 import potatoes.server.travel.dto.TravelDetailResponse;
 import potatoes.server.travel.dto.TravelPlanResponse;
-import potatoes.server.travel.dto.TravelSummaryResponse;
 import potatoes.server.travel.entity.Travel;
 import potatoes.server.travel.entity.TravelPlan;
 import potatoes.server.travel.entity.TravelUser;
@@ -40,10 +32,7 @@ import potatoes.server.travel.repository.TravelUserRepository;
 import potatoes.server.user.entity.User;
 import potatoes.server.user.repository.UserRepository;
 import potatoes.server.utils.constant.ParticipantRole;
-import potatoes.server.utils.constant.TravelSortType;
-import potatoes.server.utils.constant.TravelStatus;
 import potatoes.server.utils.error.exception.WeGoException;
-import potatoes.server.utils.pagination.dto.PageResponse;
 
 @Slf4j
 @RequiredArgsConstructor
@@ -163,52 +152,6 @@ public class TravelService {
 			bookmarkFlag);
 	}
 
-	public PageResponse<TravelSummaryResponse> getTravelList(
-		int page, int size, Boolean isDomestic, String startAt, String endAt,
-		TravelSortType sortOrder, String query, Optional<Long> userId
-	) {
-		Pageable pageable = createPageable(page, size, sortOrder);
-
-		Instant parsedStartAt = startAt != null ? parseYearMonthDay(startAt) : null;
-		Instant parsedEndAt = endAt != null ? parseYearMonthDay(endAt) : null;
-
-		Page<Travel> travels = travelRepository.findTravels(
-			isDomestic,
-			parsedStartAt,
-			parsedEndAt,
-			query,
-			pageable
-		);
-
-		Page<TravelSummaryResponse> responsePage = travels.map(travel -> {
-			int currentTravelMateCount = (int)travelUserRepository.countByTravel(travel);
-			Boolean isBookmark = userId
-				.map(uid -> bookmarkRepository.existsByUserIdAndTravelId(uid, travel.getId()))
-				.orElse(null);
-			return TravelSummaryResponse.from(travel, currentTravelMateCount, isBookmark);
-		});
-
-		return PageResponse.from(responsePage);
-	}
-
-	private Pageable createPageable(int page, int size, TravelSortType sortOrder) {
-		Sort sort = Sort.by(
-			sortOrder == TravelSortType.registrationEnd
-				? Sort.Direction.ASC
-				: Sort.Direction.DESC,
-			getSortField(sortOrder)
-		);
-		return PageRequest.of(page, size, sort);
-	}
-
-	private String getSortField(TravelSortType sortOrder) {
-		return switch (sortOrder) {
-			// TODO - 조회수를 저장하는 로직이 없어 popular조건을 추가해야됨
-			case recent, popular -> "createdAt";
-			case registrationEnd -> "registrationEnd";
-		};
-	}
-
 	public List<SimpleTravelResponse> getPopularTravels(Optional<Long> userId) {
 		return travelRepository.findTop8ByOrderByIdDesc()
 			.stream()
@@ -222,62 +165,6 @@ public class TravelService {
 				return SimpleTravelResponse.from(travel, currentTravelMate, isBookmark);
 			})
 			.collect(Collectors.toList());
-	}
-
-	@Transactional
-	public void addBookmark(Long userId, Long travelId) {
-		User user = userRepository.findById(userId).orElseThrow(() -> new WeGoException(USER_NOT_FOUND));
-		Travel travel = travelRepository.findById(travelId).orElseThrow(() -> new WeGoException(TRAVEL_NOT_FOUND));
-		if (bookmarkRepository.existsByUserAndTravel(user, travel)) {
-			return;
-		}
-
-		Bookmark bookmark = Bookmark.builder()
-			.user(user)
-			.travel(travel)
-			.build();
-		bookmarkRepository.save(bookmark);
-	}
-
-	@Transactional
-	public void deleteBookmark(Long userId, Long travelId) {
-		User user = userRepository.findById(userId).orElseThrow(() -> new WeGoException(USER_NOT_FOUND));
-		Travel travel = travelRepository.findById(travelId).orElseThrow(() -> new WeGoException(TRAVEL_NOT_FOUND));
-
-		Bookmark bookmark = bookmarkRepository.findByUserAndTravel(user, travel)
-			.orElseThrow(() -> new WeGoException(BOOKMARK_NOT_FOUND));
-		bookmarkRepository.delete(bookmark);
-	}
-
-	public PageResponse<GetMyTravelResponse> getMyTravels(int page, int size, Long userId) {
-		PageRequest request = PageRequest.of(page, size);
-		Page<GetMyTravelResponse> findTravels = travelUserRepository.findMyTravels(request, userId);
-		return PageResponse.from(findTravels);
-	}
-
-	public PageResponse<GetMyTravelResponse> getTravelsByStatus(
-		int page, int size, Long userId, TravelStatus travelStatus
-	) {
-		PageRequest request = PageRequest.of(page, size);
-		Page<GetMyTravelResponse> findTravels = travelUserRepository.findTravelsByStatus(request, userId,
-			travelStatus.name());
-		return PageResponse.from(findTravels);
-	}
-
-	public PageResponse<GetMyTravelResponse> getMyTravelsByBookmark(
-		int page, int size, Long userId
-	) {
-		PageRequest request = PageRequest.of(page, size);
-		Page<GetMyTravelResponse> findTravels = bookmarkRepository.findMyTravelsByBookmark(request, userId);
-		return PageResponse.from(findTravels);
-	}
-
-	public PageResponse<GetMyTravelResponse> getReviewableMyTravels(
-		int page, int size, Long userId
-	) {
-		PageRequest request = PageRequest.of(page, size);
-		Page<GetMyTravelResponse> findTravels = travelUserRepository.findReviewableTravels(request, userId);
-		return PageResponse.from(findTravels);
 	}
 
 	@Transactional
@@ -326,4 +213,28 @@ public class TravelService {
 		travelUserRepository.delete(travelUser);
 	}
 
+	@Transactional
+	public void addBookmark(Long userId, Long travelId) {
+		User user = userRepository.findById(userId).orElseThrow(() -> new WeGoException(USER_NOT_FOUND));
+		Travel travel = travelRepository.findById(travelId).orElseThrow(() -> new WeGoException(TRAVEL_NOT_FOUND));
+		if (bookmarkRepository.existsByUserAndTravel(user, travel)) {
+			return;
+		}
+
+		Bookmark bookmark = Bookmark.builder()
+			.user(user)
+			.travel(travel)
+			.build();
+		bookmarkRepository.save(bookmark);
+	}
+
+	@Transactional
+	public void deleteBookmark(Long userId, Long travelId) {
+		User user = userRepository.findById(userId).orElseThrow(() -> new WeGoException(USER_NOT_FOUND));
+		Travel travel = travelRepository.findById(travelId).orElseThrow(() -> new WeGoException(TRAVEL_NOT_FOUND));
+
+		Bookmark bookmark = bookmarkRepository.findByUserAndTravel(user, travel)
+			.orElseThrow(() -> new WeGoException(BOOKMARK_NOT_FOUND));
+		bookmarkRepository.delete(bookmark);
+	}
 }
